STYLE: remove all, ready to start from beginning
because the old module was not developed from python of module template, i need to start fro the begining.
But since the s4ext file pointing to this git repo, i still need to use this. so i remove everything and add the newly generated template module
then add the real code based on that.